requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
decorator==4.4.2
Django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-livereload-server==0.5.1
django-livesync==0.5
django-sslserver==0.22
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
kombu==5.3.5
mongoengine==0.27.0
moviepy==1.0.3
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
pillow==10.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymongo==3.12.1
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
websocket-client==1.7.0
whitenoise==6.6.0